Remove StatusRiotApi and create status endpoint
